Switch to extracted desktop file library
