+++ src/CMakeLists.txt
@@ -10,7 +10,7 @@ if(USE_SYSTEM_BOOST)
     set(BOOST_LIBS Boost::filesystem Boost::regex)
 else()
     # use custom built libs
-    set(BOOST_LIBS boost_system boost_filesystem boost_regex boost_lexical_cast)
+    set(BOOST_LIBS boost_system boost_filesystem boost_regex)
 endif()
